+++ .gitignore
@@ -1,4 +1,4 @@
 .nyc_output
 node_modules
-example/*/out
-example/*/tmp
+example/out
+example/tmp
