+++ example/build.pureHtml.test.js
@@ -1,5 +1,5 @@
 const tap = require('tap');
-const {build, pageMap, destDir} = require('./build.js');
+const {build, pageMap, destDir} = require('./build.pureHtml.js');
 const fs = require('fs');
 const path = require('path');
 
@@ -16,7 +16,9 @@ const cleanOut = () =>
 tap.test('pureHtml example works', ({match, notMatch}) =>
   cleanOut()
     .then(build)
-    .then(() => fs.promises.readFile(path.join(destDir, pageMap[0].dest), 'utf-8'))
+    .then(() =>
+      fs.promises.readFile(path.join(destDir, pageMap[0].dest), 'utf-8'),
+    )
     .then(renderedHtml => {
       // console.log(renderedHtml)
       match(renderedHtml, /<footer/, 'contains footer');
