+++ example/src/Foot.svelte
