+++ example/src/Page.svelte
